backend/neural_network/requirements.txt
fastapi
accelerate>=0.16.0
diffusers
torchvision
transformers>=4.25.1
numpy
opencv-contrib-python
Pillow
requests